pyproject.toml
[project]
name = "bidauto-web-orders"
version = "0.1.0"
description = ""
authors = [
    {name = "ivanpetrovskij",email = "peyrovskaaa@gmai.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.120.0,<0.121.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "fastapi-problem (>=0.11.6,<0.12.0)",
    "redis (>=7.0.1,<8.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "grpcio (>=1.76.0,<2.0.0)",
    "authtools @ git+https://github.com/vinasLT/shared_auth.git@v0.1.8",
    "protobuf (>=6.33.1,<7.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "fastapi-pagination (>=0.15.0,<0.16.0)",
    "reportlab (>=4.4.5,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "aiosqlite (>=0.21.0,<0.22.0)"
]
